Added coolant control (M7*, M8, M9). Mist control can be enabled via config.h.
- Added coolant control! Flood control (M8) functions on analog pin 0.
Mist control (M7) is compile-time optional and is on analog pin 1. (Use
only if you have multiple coolants on your system). Based on work by
@openpnp.

- Fixed some variable assignments in spindle control.

+++ spindle_control.c
@@ -43,7 +43,7 @@ void spindle_stop()
   SPINDLE_ENABLE_PORT &= ~(1<<SPINDLE_ENABLE_BIT);
 }
 
-void spindle_run(int direction, uint32_t rpm) 
+void spindle_run(int8_t direction, uint16_t rpm) 
 {
   if (direction != current_direction) {
     plan_synchronize();
